pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Marc Streeter <mstreeter@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.73.0"
uvicorn = {extras = ["standard"], version = "^0.17.4"}
typer = "^0.4.0"
tortoise-orm = {extras = ["aiomysql"], version = "^0.18.1"}
orjson = "^3.6.6"

[tool.poetry.scripts]
jane = "cli:app"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
